Fixed bug

+++ appFBLA2019/appFBLA2019/LevelEditor/CreateNewLevelPage.xaml.cs
@@ -202,7 +202,7 @@ namespace appFBLA2019
                 }
 
                 // If they renamed the level, delete the old one
-                if (this.originalName != this.EntryLevelName.Text.Trim())
+                if (this.originalName != this.EntryLevelName.Text.Trim() && this.originalAuthor == CredentialManager.Username)
                 {
                     Directory.Delete(App.Path + "/" + this.originalName + "`" + this.originalAuthor, true);
                 }
